# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: MSS Platform API
  description: |
    MSS Platform API provides comprehensive endpoints for managing cybersecurity services,
    multi-tenant organizations, user management, and security operations.
    
    ## Authentication
    
    The API uses JWT Bearer tokens for authentication. Most endpoints require a valid JWT token
    in the Authorization header:
    
    ```
    Authorization: Bearer <jwt_token>
    ```
    
    ## Multi-Factor Authentication Flow
    
    1. **Login**: POST /api/auth/login with credentials
       - Returns `mfa_token` if MFA is enabled
       - Returns `access_token` and `refresh_token` if MFA is disabled
    
    2. **Complete MFA**: POST /api/auth/mfa/complete with MFA token and TOTP/backup code
       - Returns `access_token` and `refresh_token`
    
    3. **Refresh Token**: POST /api/auth/refresh when access token expires
    
    ## Role-Based Access Control
    
    ### Customer Organization Roles
    - **admin**: Full organization management
    - **report_viewer**: Read access to security reports
    - **request_user**: Can request interventions
    - **basic_user**: Limited dashboard access
    
    ### MSS Provider Roles
    - **super_admin**: Platform-wide administration
    - **account_manager**: Customer organization management
    - **security_analyst**: Security monitoring and analysis
    - **technician**: Customer environment access
    
    ## Rate Limiting
    
...
    
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token obtained from login/MFA completion
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        message:
          type: string
          description: Human-readable message
        error:
          type: string
          description: Error message if success is false
      required:
        - success
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
          enum: [admin, report_viewer, request_user, basic_user, super_admin, technician, security_analyst, account_manager]
        status:
          type: string
          enum: [active, suspended, locked, deleted]
        mfaEnabled:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
          nullable: true
        lastLoginIp:
          type: string
          nullable: true
        failedLoginAttempts:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        organization:
          $ref: '#/components/schemas/Organization'
      required:
        - id
        - email
        - firstName
        - lastName
        - role
        - status
        - mfaEnabled
        - createdAt
        - updatedAt
    CreateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        role:
          type: string
          enum: [admin, report_viewer, request_user, basic_user, technician, security_analyst, account_manager]
        orgId:
          type: string
          format: uuid
          description: Target organization ID (optional - defaults to current user's organization)
        password:
          type: string
          minLength: 8
          description: Optional password (if not provided, user will receive invitation email)
        sendInvitation:
          type: boolean
          default: true
          description: Whether to send invitation email
      required:
        - email
        - firstName
        - lastName
        - role
    UpdateUserRequest:
      type: object
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        email:
          type: string
          format: email
          description: Email updates require MSS provider role
        role:
          type: string
          enum: [admin, report_viewer, request_user, basic_user, technician, security_analyst, account_manager]
          description: Cannot change own role
        status:
          type: string
          enum: [active, suspended, locked]
          description: Status updates require admin privileges

    # Organization schemas
...
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            success:
              type: boolean
              example: false
            error:
              type: string
              description: Error message
            code:
              type: string
              description: Error code for programmatic handling
              example: VALIDATION_ERROR
            details:
              type: object
              additionalProperties: true
              description: Additional error details
          required:
            - success
            - error
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Authentication required"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Insufficient permissions"
            code: "FORBIDDEN"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Resource not found"
            code: "NOT_FOUND"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              email: "Invalid email format"
              password: "Password must be at least 8 characters"
...
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users
      description: |
        Get list of users with organization-based access control:
        - Customer admins: See only users in their organization
        - MSS provider users: See customer users or their own org users
        - Super admins: See all users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: orgId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by organization ID
        - name: role
          in: query
          schema:
            type: string
            enum: [admin, report_viewer, request_user, basic_user, super_admin, technician, security_analyst, account_manager]
          description: Filter by role
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended, locked]
          description: Filter by status (excludes deleted users)
        - name: search
          in: query
          schema:
            type: string
          description: Search in first name, last name, or email
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          users:
                            type: array
                            items:
                              $ref: '#/components/schemas/User'
                          total:
                            type: integer
                          page:
                            type: integer
                          limit:
                            type: integer
                          totalPages:
                            type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      tags:
        - Users
      summary: Create new user
      description: |
        Create a new user in the organization.
        Authorization:
        - Customer admins: Create users in their own organization only
        - MSS account managers/super admins: Create users in any organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
                      message:
                        type: string
                        example: "User created and invitation sent"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Organization not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: User with email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /users/profile:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: Get the profile information of the currently authenticated user
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /users/{userId}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: |
        Get specific user information.
        Authorization:
        - Users can view their own profile
        - Customer admins can view users in their organization
        - MSS managers can view customer users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - Users
      summary: Update user
      description: |
        Update user information.
        Authorization:
        - Users can update their own basic info (not role/status)
        - Customer admins can update users in their organization
        - MSS managers can update customer users
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          id:
                            type: string
                          email:
                            type: string
                          firstName:
                            type: string
                          lastName:
                            type: string
                          role:
                            type: string
                          status:
                            type: string
                          updatedAt:
                            type: string
                            format: date-time
                      message:
                        type: string
                        example: "User updated successfully"
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - Users
      summary: Delete user
      description: |
        Soft delete user (marks as deleted and revokes sessions).
        Authorization:
        - Customer admins can delete users in their organization (except themselves)
        - MSS super admins can delete any user (except themselves)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      message:
                        type: string
                        example: "User deleted successfully"
        '400':
          description: Cannot delete own account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # Organization management endpoints
...